% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{D3 Manhattan Plot Visualizer}
\usage{
manhattanPlot(dataset, width = NULL, height = NULL, snps_col = "SNP",
  pvals_col = "PVal", sigLine = TRUE, animationSpeed = 1000,
  logged = FALSE)
}
\arguments{
\item{dataset}{Your data. Contains SNP names and PValues as columns of a dataframe.}

\item{width}{Set custom width.}

\item{height}{Set custom height.}

\item{snps_col}{The name of the column of your dataframe containing the SNP names. Defaults to "SNP".}

\item{pvals_col}{The name of the column of your dataframe containing the PValues. Defaults to "PVal".}

\item{sigLine}{Draw a line indicating significance threshold according to a conservative bonferroni correction.}

\item{animationSpeed}{How fast you want the program to animate in milliseconds. To turn off set to 0.}

\item{logged}{Have the PValues already been -log10()'d? Defaults to false.}
}
\description{
This generates an interactive manhattan plot using d3.js.
Mouse over snps to see name and pvalue. Now you don't have to waste your time
writing R code to find significant snp identities.
}
\examples{
#Load sample dataset from the package.
d = sampleVals
#and plot. It's that easy.
manhattanPlot(d, sigLine = FALSE)
}
\keyword{GWAS}
\keyword{genetics,}
\keyword{manhattan}
\keyword{plot,}
\keyword{statistical}

