% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\docType{data}
\name{random}
\alias{random}
\title{Randomly generated snps and pvals}
\format{A data frame with 100 rows and 2 columns:
\describe{
  \item{SNP}{Snp names. Character}
  \item{PVal}{P-Values. Float}
}}
\usage{
random
}
\description{
Randomly generated dataset of 400 snps. One should result as significant using
the bonferroni correction threshold built in.
}
\keyword{datasets}

