% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{renderManhattanPlot}
\alias{renderManhattanPlot}
\title{Widget render function for use in Shiny}
\usage{
renderManhattanPlot(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

