% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\docType{data}
\name{sampleVals}
\alias{sampleVals}
\title{Sample SNPS and Pvals from Klein et al. 2004}
\format{A data frame with 100 rows and 2 columns:
\describe{
  \item{SNP}{Snp names. Character}
  \item{PVal}{-log10 of the pvalues. Float}
}}
\source{
Klein, R. J., Zeiss, C., Chew, E. Y., Tsai, J. Y., Sackler, R. S., Haynes,
C., ... & Hoh, J. (2005). Complement factor H polymorphism in age-related
macular degeneration. Science, 308(5720), 385-389.
}
\usage{
sampleVals
}
\description{
This is a subset of the data from the 2005 paper in Science
"Complement Factor H Polymorphism in Age-Related Macular Degeneration". The
data are subsetted to the 100 most significant snps when stratifying by the
male gender.
}
\keyword{datasets}

